restructure and rename of vulnerability form
