+++ src/components/Vulnerability/FormSteps/ContactInfo/ContactInfo.tsx
@@ -1,7 +1,7 @@
 import { isAddress } from "ethers/lib/utils";
 import { useCallback, useContext, useEffect, useState } from "react";
-import "../../styles/Vulnerability/ContactInfo.scss";
-import { AccordionContext } from "./VulnerabilityAccordion";
+import "styles/Vulnerability/ContactInfo.scss";
+import { AccordionContext } from "../../VulnerabilityForm";
 import classNames from "classnames";
 import { useEthers } from "@usedapp/core";
