+++ src/components/Vulnerability/FormSteps/ProjectSelect/ProjectSelect.tsx
@@ -1,11 +1,11 @@
 import { useContext, useState } from "react";
 import { useSelector } from "react-redux";
-import { RootState } from "../../../reducers";
-import SearchIcon from "../../../assets/icons/search.icon";
-import { IVault } from "../../../types/types";
-import Loading from "../../Shared/Loading";
-import { AccordionContext } from "../VulnerabilityAccordion";
-import { ScreenSize } from "../../../constants/constants";
+import { RootState } from "../../../../reducers";
+import SearchIcon from "../../../../assets/icons/search.icon";
+import { IVault } from "../../../../types/types";
+import Loading from "../../../Shared/Loading";
+import { AccordionContext } from "../../VulnerabilityForm";
+import { ScreenSize } from "../../../../constants/constants";
 import { useVaults } from "hooks/useVaults";
 import Project from "./components/Project/Project";
 import "./index.scss";
