+++ src/components/Vulnerability/FormSteps/ProjectSelect/components/Project/index.scss
@@ -1,4 +1,4 @@
-@import "../../../../../styles/constants.scss";
+@import "../../../../../../styles/constants.scss";
 
 .project-row {
   cursor: pointer;
