+++ src/components/Vulnerability/FormSteps/ProjectSelect/index.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/constants.scss";
+@import "../../../../styles/constants.scss";
 
 .project-select-wrapper {
   position: relative;
