+++ src/components/Vulnerability/FormSteps/Submit/Submit.tsx
@@ -2,7 +2,7 @@ import SubmitSuccess from "./components/SubmitSuccess/SubmitSuccess";
 import SubmitReview from "./components/SubmitReview/SubmitReview";
 import { useContext } from "react";
 import "./index.scss";
-import { AccordionContext } from "../VulnerabilityAccordion";
+import { AccordionContext } from "../../VulnerabilityForm";
 
 /**
  * Function to call the Telegram bot after a success on the blockchain transaction.
