+++ src/components/Vulnerability/FormSteps/Submit/components/SubmitReview/SubmitReview.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { AccordionContext } from "../../../VulnerabilityAccordion";
+import { AccordionContext } from "../../../../VulnerabilityForm";
 import Loading from "components/Shared/Loading";
 import EditIcon from "assets/icons/edit.svg";
 import "./index.scss";
