+++ src/components/Vulnerability/FormSteps/Submit/components/SubmitReview/index.scss
@@ -1,4 +1,4 @@
-@import "../../../../../styles/constants.scss";
+@import "../../../../../../styles/constants.scss";
 
 .submit-review-wrapper {
   display: flex;
