+++ src/components/Vulnerability/FormSteps/Submit/components/SubmitSuccess/SubmitSuccess.tsx
@@ -1,10 +1,10 @@
 
 import { useContext } from "react";
 import { useTranslation } from "react-i18next";
-import DiscordIcon from "../../../../../assets/icons/social/discord.icon";
-import { Colors, SocialLinks } from "../../../../../constants/constants";
-import Dot from "../../../../Shared/Dot/Dot";
-import { AccordionContext } from "../../../VulnerabilityAccordion";
+import DiscordIcon from "../../../../../../assets/icons/social/discord.icon";
+import { Colors, SocialLinks } from "../../../../../../constants/constants";
+import Dot from "../../../../../Shared/Dot/Dot";
+import { AccordionContext } from "../../../../VulnerabilityForm";
 import "./index.scss";
 
 export enum BotStatus {
