+++ src/components/Vulnerability/FormSteps/Submit/components/SubmitSuccess/index.scss
@@ -1,4 +1,4 @@
-@import "../../../../../styles/constants.scss";
+@import "../../../../../../styles/constants.scss";
 
 .submit-success-wrapper {
   .transaction-summary-wrapper {
