+++ src/components/Vulnerability/FormSteps/Submit/index.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/constants.scss";
+@import "../../../../styles/constants.scss";
 
 .submit-wrapper {
   position: relative;
