+++ src/components/Vulnerability/FormSteps/TermsAndProcess/TermsAndProcess.tsx
@@ -1,12 +1,12 @@
 import { useContext, useEffect, useState } from "react";
-import { AccordionContext } from "../VulnerabilityAccordion";
+import { AccordionContext } from "../../VulnerabilityForm";
 import { useTranslation } from "react-i18next";
-import { TERMS_OF_USE } from "../../../constants/constants";
-import QuestionIcon from "../../../assets/icons/question-icon.svg";
-import ErrosIcon from "../../../assets/icons/error-icon.svg";
-import TimelineIcon from "../../../assets/icons/timeline-icon.svg";
-import BugIcon from "../../../assets/icons/bug-icon.svg";
-import RewardIcon from "../../../assets/icons/reward-icon.svg";
+import { TERMS_OF_USE } from "constants/constants";
+import QuestionIcon from "assets/icons/question-icon.svg";
+import ErrosIcon from "assets/icons/error-icon.svg";
+import TimelineIcon from "assets/icons/timeline-icon.svg";
+import BugIcon from "assets/icons/bug-icon.svg";
+import RewardIcon from "assets/icons/reward-icon.svg";
 import Award from "./components/Award/Award";
 import "./index.scss";
 import { useVaults } from "hooks/useVaults";
