+++ src/components/Vulnerability/FormSteps/TermsAndProcess/index.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/constants.scss";
+@import "../../../../styles/constants.scss";
 
 .terms-and-process-wrapper {
   color: $white;
