+++ src/components/Vulnerability/FormSteps/VulnerabilityDescription/VulnerabilityDescription.tsx
@@ -1,11 +1,11 @@
 import { useCallback, useContext, useEffect, useState } from "react";
-import "../../styles/Vulnerability/Description.scss";
-import { AccordionContext } from "./VulnerabilityAccordion";
+import "styles/Vulnerability/Description.scss";
+import { AccordionContext } from "../../VulnerabilityForm";
 import MDEditor from '@uiw/react-md-editor';
 import download from "downloadjs";
 import classNames from "classnames";
 import { Colors } from "constants/constants";
-import { encryptWithKeys } from "./encrypt";
+import { encryptWithKeys } from "../../encrypt";
 import { useVaults } from "hooks/useVaults";
 
 export default function Description() {
