+++ src/components/Vulnerability/VulnerabilityForm.tsx
@@ -1,17 +1,17 @@
 import { createContext, Dispatch, SetStateAction, useCallback, useEffect, useState } from "react";
 import "../../styles/Vulnerability/VulnerabilityAccordion.scss";
-import AccordionCard from "./AccordionCard";
-import ProjectSelect from "./ProjectSelect/ProjectSelect";
-import Description from "./Description";
-import ContactInfo from "./ContactInfo";
-import TermsAndProcess from "./TermsAndProcess/TermsAndProcess";
-import Submit from "./Submit/Submit";
+import VulnerabilityFormCard from "./VulnerabilityFormCard";
+import ProjectSelect from "./FormSteps/ProjectSelect/ProjectSelect";
+import Description from "./FormSteps/VulnerabilityDescription/VulnerabilityDescription";
+import ContactInfo from "./FormSteps/ContactInfo/ContactInfo";
+import TermsAndProcess from "./FormSteps/TermsAndProcess/TermsAndProcess";
+import Submit from "./FormSteps/Submit/Submit";
 import { getAppVersion } from "../../utils";
 import { LocalStorage } from "../../constants/constants";
 import { Card, IVulnerabilityData } from "./types";
 import { useEthers, useTransactions } from "@usedapp/core";
 import { useVaults } from "hooks/useVaults";
-import { BotStatus } from "./Submit/components/SubmitSuccess/SubmitSuccess";
+import { BotStatus } from "./FormSteps/Submit/components/SubmitSuccess/SubmitSuccess";
 import axios from "axios";
 import { DEFAULT_BOT } from "settings";
 import { useClaim } from "hooks/contractHooks";
@@ -45,7 +45,7 @@ export const VULNERABILITY_INIT_DATA: IVulnerabilityData = {
 
 
 // updateStorage(singleVulnerability): TODO: 
-export default function VulnerabilityAccordion() {
+export default function VulnerabilityForm() {
   const submittingVurnability = useTransactions().transactions.some(tx => !tx.receipt)
   const [currentStep, setCurrentStep] = useState<number>();
   const [cards, setCards] = useState<IVulnerabilityData>();
@@ -142,7 +142,7 @@ export default function VulnerabilityAccordion() {
       <div id="accrodionWrapper" className="accrodion-wrapper">
         <AccordionContext.Provider value={context}>
           {steps.map((step, index) => (
-            <AccordionCard
+            <VulnerabilityFormCard
               extraInfoTitle={index === 0 ? cards?.project?.projectName : ""}
               id={index}
               key={index}
@@ -152,7 +152,7 @@ export default function VulnerabilityAccordion() {
               verified={cards?.[Card[step.card]] !== undefined}
               disabled={submittingVurnability} >
               {<step.component />}
-            </AccordionCard>))}
+            </VulnerabilityFormCard>))}
         </AccordionContext.Provider>
       </div>
     </div>
