+++ src/components/Vulnerability/VulnerabilityFormCard.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import classNames from "classnames";
-import { AccordionContext } from "./VulnerabilityAccordion";
+import { AccordionContext } from "./VulnerabilityForm";
 import ArrowIcon from "../../assets/icons/arrow.icon";
 import CheckmarkIcon from "../../assets/icons/checkmark.icon";
 import "../../styles/Vulnerability/AccordionCard.scss";
@@ -18,7 +18,7 @@ interface IProps {
   collapsed: boolean
 }
 
-export default function AccordionCard({ id, card, title, verified, children, disabled, extraInfoTitle, collapsed }: IProps) {
+export default function VulnerabilityFormCard({ id, card, title, verified, children, disabled, extraInfoTitle, collapsed }: IProps) {
   const { setCurrentStep } = useContext(AccordionContext);
   const cardContentClass = classNames({
     "card-content-wrapper": true,
