+++ src/components/Vulnerability/types.ts
@@ -1,5 +1,5 @@
 import { SessionKey } from "openpgp";
-import { BotStatus } from "./Submit/components/SubmitSuccess/SubmitSuccess";
+import { BotStatus } from "./FormSteps/Submit/components/SubmitSuccess/SubmitSuccess";
 
 export enum Card {
   project,
